Add files via upload
Data Cleaning in SQL Server. This project looks to analyze the Top 10 regions affected by COVID 19 in India. The aim is to identify the primary regions to dispose vaccinations. This section is focused on cleaning the data to ensure that it is ready for the Tableau section. To review the Tableau section, please review the following link:

https://public.tableau.com/app/profile/james.savage4094/viz/Top10RegionswithCOVID19inIndia/Dashboard1